Refactor project creation
